ENH: Improve fit for PCA and Labelmap_to_Labelmap registration (#69)
* ENH: Improve fit for PCA and Labelmap_to_Labelmap registration

+++ src/physiomotion4d/contour_tools.py
@@ -264,7 +264,7 @@ class ContourTools(PhysioMotion4DBase):
                 reference_image.GetLargestPossibleRegion().GetSize()[1],
                 reference_image.GetLargestPossibleRegion().GetSize()[0],
             ),
-            dtype=np.uint8,
+            dtype=np.uint16,
         )
         for i, mesh in enumerate(meshes):
             mask_image = self.create_mask_from_mesh(mesh, reference_image)
